refactor: improve folder/package structure
closes #109
