+++ src/main/java/ci/controller/CommitController.java
@@ -1,5 +1,6 @@
-package ci;
+package ci.controller;
 
+import ci.service.Status;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
