+++ src/main/java/ci/service/CompilationService.java
@@ -1,4 +1,4 @@
-package ci;
+package ci.service;
 
 import java.io.BufferedReader;
 import java.io.File;
