+++ src/main/java/ci/util/RepoSetup.java
@@ -1,4 +1,4 @@
-package ci;
+package ci.util;
 
 import java.io.BufferedReader;
 import java.io.File;
