+++ src/test/java/ci/integration/GithubAPIHandlerTest.java
@@ -3,6 +3,7 @@ package ci;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import ci.integration.GithubAPIHandler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.util.Random;
