+++ src/test/java/ci/service/CompilationTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import ci.service.CompilationService;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
