+++ src/test/java/ci/service/StatusTest.java
@@ -4,7 +4,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import ci.Status.CommitRecord;
+import ci.service.Status;
+import ci.service.Status.CommitRecord;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
