+++ src/test/java/ci/util/RepoSetupTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import ci.util.RepoSetup;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
